LPS-145881 Reorganize SXPIndexerPostProcessor to be possible index contentLength and versionCount in Portal CE
