+++ modules/apps/journal/journal-service/src/main/java/com/liferay/journal/internal/search/JournalArticleIndexer.java
@@ -420,6 +420,9 @@ public class JournalArticleIndexer extends BaseIndexer<JournalArticle> {
 			StringUtil.split(journalArticle.getTreePath(), CharPool.SLASH));
 		document.addKeyword(Field.VERSION, journalArticle.getVersion());
 
+		document.addNumber(
+			"versionCount", GetterUtil.getDouble(journalArticle.getVersion()));
+
 		document.addKeyword(
 			"ddmStructureKey", journalArticle.getDDMStructureKey());
 		document.addKeyword(
